Dashboard is looking better, multiple visual page fixes, cleaned up and removed uneeded components
