+++ client/src/components/NavBar.jsx
@@ -18,68 +18,21 @@ const NavBar = () => {
               </Link>
             </li>
             <li>
-              <button
-                id="dropdownNavbarLink"
-                data-dropdown-toggle="dropdownNavbar"
+              <Link
+                to="/createnewpoll"
                 className="flex w-full items-center justify-between border-b border-gray-100 py-2 pr-4 pl-3 text-gray-700 hover:bg-gray-50 dark:border-gray-700 dark:text-gray-300 dark:hover:bg-gray-700 dark:hover:text-white dark:focus:text-white md:w-auto md:border-0 md:p-0 md:hover:bg-transparent md:hover:text-blue-700 md:dark:hover:bg-transparent"
               >
-                Polls{" "}
-                <svg
-                  className="ml-1 h-4 w-4"
-                  aria-hidden="true"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                >
-                  <path
-                    fillRule="evenodd"
-                    d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
-                    clipRule="evenodd"
-                  ></path>
-                </svg>
-              </button>
-
-              {/*POLLS DROP DOWN MENU */}
-              {/*I FIGURED POLLS WOULD BE BEST SUITED IN A DROPDOWN TO SAVE ON SPACE. CREATE POLL, VIEW POLL, ECT.  */}
-              <div
-                id="dropdownNavbar"
-                className="z-10 hidden w-44 divide-y divide-gray-100 rounded bg-slate-50 shadow dark:divide-gray-600 dark:bg-gray-700"
-              >
-                {/*DROP DOWN LIST*/}
-                <ul
-                  className="py-1 text-sm text-gray-700 dark:text-gray-400"
-                  aria-labelledby="dropdownLargeButton"
-                >
-                  {/*VIEW POLL*/}
-                  <li>
-                    <a href="#" className="block py-2 px-4">
-                      View Polls
-                    </a>
-                  </li>
-
-                  {/*CREATE NEW POLL*/}
-                  <li>
-                    <a href="#" className="block py-2 px-4">
-                      Create New Poll
-                    </a>
-                  </li>
-
-                  {/*DELETE POLL....ECT */}
-                  <li>
-                    <a href="#" className="block py-2 px-4">
-                      Delete Poll
-                    </a>
-                  </li>
-                </ul>
-              </div>
+                Create New Poll
+              </Link>
             </li>
 
             {/*Link 1*/}
             <li>
               <Link
-                to="/createnewpoll"
+                to="/poll"
                 className="block rounded py-2 pr-4 pl-3 text-gray-700 hover:bg-gray-100 dark:text-gray-300 dark:hover:bg-gray-700 dark:hover:text-white md:border-0 md:p-0 md:hover:bg-transparent md:hover:text-blue-700 md:dark:hover:bg-transparent md:dark:hover:text-white"
               >
-                Create New Poll
+                Open Polls
               </Link>
             </li>
